[Rename] `StreamSubscriber` -->  `EventSubscriber`
**Major**:
- Project-wide symbol renaming
- Also renamed:
  - `*stream_subscriber*` to `*event_subscriber*`
  - `*_stream_subscribers*` to `*_stream_subscribers_examples*`
