+++ hatchling/mcp_utils/mcp_tool_call_subscriber.py
@@ -7,14 +7,14 @@ and dispatches them to MCPToolExecution for processing.
 from collections import deque
 from json import dumps as json_dumps
 
-from hatchling.core.llm.streaming_management import StreamSubscriber, StreamEvent, StreamEventType
+from hatchling.core.llm.streaming_management import EventSubscriber, StreamEvent, StreamEventType
 from hatchling.core.llm.data_structures import ToolCallParsedResult
 from hatchling.core.llm.providers.registry import ProviderRegistry
 from hatchling.core.logging.logging_manager import logging_manager
 from .mcp_tool_execution import MCPToolExecution
 
 
-class MCPToolCallSubscriber(StreamSubscriber):
+class MCPToolCallSubscriber(EventSubscriber):
     """Subscriber that handles LLM_TOOL_CALL_REQUEST events and dispatches them for execution.
 
     This subscriber only processes one tool call per request ID, using a rolling buffer
